lots of misc improvements
A non-exhaustive list of things I did:
* reduce warnings
* delete useless block files (mc_bricks and mc_sponge)
* change #ifdef formatting to better match the convention
* make deleted `operator=` declarations consistent
* delete (some) unneeded includes
* add (some) needed includes
* replace Util::change_directory with std::filesystem::current_path

+++ src/position/BlockInChunk.hpp
@@ -27,22 +27,22 @@ struct BlockInChunk
 
 	value_type operator[](const std::ptrdiff_t i) const
 	{
-		#ifdef DEBUG_BUILD
+	#ifdef DEBUG_BUILD
 		if(i > 2)
 		{
 			throw std::out_of_range("Position::BlockInChunk::operator[]: " + std::to_string(i) + " > 2");
 		}
-		#endif
+	#endif
 		return (&x)[i];
 	}
 	value_type& operator[](const std::ptrdiff_t i)
 	{
-		#ifdef DEBUG_BUILD
+	#ifdef DEBUG_BUILD
 		if(i > 2)
 		{
 			throw std::out_of_range("Position::BlockInChunk::operator[]: " + std::to_string(i) + " > 2");
 		}
-		#endif
+	#endif
 		return (&x)[i];
 	}
 	BlockInChunk& operator+=(const BlockInChunk&);
@@ -54,9 +54,9 @@ struct BlockInChunk
 	value_type x, y, z;
 
 private:
-	#ifdef DEBUG_BUILD
+#ifdef DEBUG_BUILD
 	void check_bounds();
-	#endif
+#endif
 };
 
 std::ostream& operator<<(std::ostream&, const BlockInChunk&);
@@ -68,7 +68,7 @@ namespace glm {
 template <typename CharT, typename Traits, precision P>
 GLM_FUNC_QUALIFIER std::basic_ostream<CharT, Traits>& operator<<(std::basic_ostream<CharT, Traits>& os, tvec3<uint8_t, P> const& a)
 {
-	tvec3<uint_fast16_t, P> b(a);
+	tvec3<unsigned int, P> b(a);
 	return detail::print_vector_on(os, b);
 }
